Split cmd/main_test.go's standalone functions into topical files
Purely mechanical: no test bodies changed. Moves everything except
TestE2EParallel (which must stay a single function for its shared
e2eClients()/parallel-pool setup) into local_test.go, e2e_vanilla_test.go,
e2e_dynamic_test.go, and e2e_helpers_test.go, so the file most people touch
for a routine change is no longer 2400+ lines.
