+++ cmd/main_test.go
@@ -1,21 +1,15 @@
 package main
 
 import (
-	"bytes"
 	"context"
-	"encoding/json"
 	"fmt"
 	"os"
 	"os/exec"
 	"path"
-	"path/filepath"
-	"regexp"
 	"strings"
 	"testing"
 	"time"
 
-	"github.com/spf13/pflag"
-	"github.com/spf13/viper"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	admissionregistrationv1 "k8s.io/api/admissionregistration/v1"
@@ -27,436 +21,9 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime/schema"
-	"k8s.io/client-go/dynamic"
-	"k8s.io/client-go/kubernetes"
-	"k8s.io/client-go/tools/clientcmd"
-	"k8s.io/client-go/util/retry"
-	"k8s.io/klog/v2"
 
 	"github.com/bergerx/kubectl-status/pkg/plugin"
 )
-
-type cmdTest struct {
-	name            string
-	args            []string
-	stdoutRegexPath string // Regex match against file contents under test folder
-	stdoutEqualPath string // Exact match with file contents under test folder
-	stderrRegex     string // Regex
-	stderrEqual     string // Exact
-	wantErr         string // Contains
-}
-
-// createBadNode creates a synthetic Node (no real kubelet backs it) that's cordoned, tainted,
-// and reporting NotReady/MemoryPressure -- everything pod_node_problems/pod_node_problem_flags
-// are meant to surface. It registers cleanup and returns the Node's name.
-func createBadNode(t *testing.T, clientset *kubernetes.Clientset) string {
-	t.Helper()
-	node := &corev1.Node{
-		ObjectMeta: metav1.ObjectMeta{
-			GenerateName: "kubectl-status-test-bad-node-",
-		},
-		Spec: corev1.NodeSpec{
-			Unschedulable: true,
-			Taints: []corev1.Taint{
-				{Key: "dedicated", Value: "gpu", Effect: corev1.TaintEffectNoSchedule},
-			},
-		},
-	}
-	node, err := clientset.CoreV1().Nodes().Create(context.TODO(), node, metav1.CreateOptions{})
-	require.NoError(t, err)
-	t.Cleanup(func() {
-		clientset.CoreV1().Nodes().Delete(context.TODO(), node.Name, metav1.DeleteOptions{})
-	})
-	// The real node-lifecycle-controller starts reconciling this Node as soon as it's created
-	// (e.g. adding its own NotReady taint), racing our status update -- retry on conflict.
-	err = retry.RetryOnConflict(retry.DefaultRetry, func() error {
-		latest, err := clientset.CoreV1().Nodes().Get(context.TODO(), node.Name, metav1.GetOptions{})
-		if err != nil {
-			return err
-		}
-		latest.Status.Conditions = []corev1.NodeCondition{
-			{
-				Type:               corev1.NodeReady,
-				Status:             corev1.ConditionFalse,
-				Reason:             "KubeletNotReady",
-				Message:            "kubelet is not ready",
-				LastTransitionTime: metav1.Now(),
-			},
-			{
-				Type:               corev1.NodeMemoryPressure,
-				Status:             corev1.ConditionTrue,
-				Reason:             "KubeletHasInsufficientMemory",
-				Message:            "kubelet has insufficient memory available",
-				LastTransitionTime: metav1.Now(),
-			},
-		}
-		_, err = clientset.CoreV1().Nodes().UpdateStatus(context.TODO(), latest, metav1.UpdateOptions{})
-		return err
-	})
-	require.NoError(t, err)
-	return node.Name
-}
-
-func nodeNameModifier(stdout string) string {
-	return string(regexp.MustCompile(`Node/[a-z0-9-]+`).ReplaceAll([]byte(stdout), []byte(`Node/minikube`)))
-}
-
-func (c cmdTest) assert(t *testing.T, stdoutModifier func(string) string, opts ...func(*plugin.RenderConfig)) {
-	t.Helper()
-	t.Logf("running cmdTest assert: %s", c)
-	stdout, stderr, err := executeCMD(t, c.args, opts...)
-	if stdoutModifier != nil {
-		stdout = nodeNameModifier(stdout)
-	}
-	switch {
-	case c.stdoutRegexPath == "" && c.stdoutEqualPath == "":
-		assert.Empty(t, stdout)
-	case c.stdoutEqualPath != "":
-		outFile := path.Join("..", "tests", c.stdoutEqualPath)
-		out, err := os.ReadFile(outFile)
-		assert.NoErrorf(t, err, "failed to read test artifact file: %s", outFile)
-		assert.Equal(t, string(out), stdout)
-	case c.stdoutRegexPath != "":
-		outFile := path.Join("..", "tests", c.stdoutRegexPath)
-		regexBytes, err := os.ReadFile(outFile)
-		assert.NoErrorf(t, err, "failed to read test artifact file: %s", outFile)
-		regex := `(?ms)` + string(regexBytes)
-		assert.Regexp(t, regex, stdout)
-	}
-	switch {
-	case c.stderrRegex == "" && c.stderrEqual == "":
-		assert.Empty(t, stderr)
-	case c.stderrRegex != "":
-		assert.Regexp(t, c.stderrRegex, stderr)
-	case c.stderrEqual != "":
-		assert.Equal(t, c.stderrEqual, stderr)
-	}
-	if c.wantErr != "" {
-		assert.ErrorContains(t, err, c.wantErr)
-	}
-}
-
-func TestRootCmdWithoutACluster(t *testing.T) {
-	t.Setenv("KUBECONFIG", "/dev/null")
-	opts := []func(*plugin.RenderConfig){
-		func(cfg *plugin.RenderConfig) {
-			cfg.DurationRound = func(_ interface{}) string { return "1m" }
-		},
-	}
-	tests := []cmdTest{
-		{
-			name:        "empty call should print an error and simple usage",
-			stderrRegex: `You must provide one or more resources by argument or filename.\nExample resource specifications`,
-		},
-		{
-			name:        "pods against a non-configured client should print an error",
-			args:        []string{"pods"},
-			stderrRegex: `the server.*(rejected|refused)`,
-		},
-		{
-			name:        "missing file should fail",
-			args:        []string{"-f", "non-existing.yaml"},
-			stderrRegex: `error: the path \"non-existing.yaml\" does not exist\n$`,
-		},
-		{
-			name:        "file without local should fail",
-			args:        []string{"-f", "../tests/artifacts/deployment-healthy.yaml"},
-			stderrRegex: `the server rejected our request for an unknown reason|connect: connection refused`,
-		},
-		{
-			name:            "file with local should succeed",
-			args:            []string{"-f", "../tests/artifacts/deployment-healthy.yaml", "--local"},
-			stdoutRegexPath: "artifacts/deployment-healthy.local.regex",
-		},
-		{
-			name:            "cr file with local with status should render ready conditions",
-			args:            []string{"-f", "../tests/artifacts/cr-dbconn-mymysql.yaml", "--local"},
-			stdoutRegexPath: "artifacts/cr-dbconn-mymysql.local.regex",
-		},
-		{
-			name:            "file with 'kind: List' should list all resources",
-			args:            []string{"-f", "../tests/artifacts/multiple-2-pods-list.yaml", "--local"},
-			stdoutRegexPath: "artifacts/multiple-2-pods-list.local.regex",
-		},
-		{
-			name:            "file with multiple yaml documents should list all resources",
-			args:            []string{"-f", "../tests/artifacts/multiple-2-pods-docs.yaml", "--local"},
-			stdoutRegexPath: "artifacts/multiple-2-pods-docs.local.regex",
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			tt.assert(t, nil, opts...)
-		})
-	}
-}
-
-// TestDeepRespectsExplicitIncludeFalse guards a fix to --deep: it used to force every
-// --include-* flag to true unconditionally (enableAllIncludes), silently overriding an explicit
-// --include-events=false/--include-managed-fields=false passed alongside it on the same command
-// line -- exactly the flags the e2e "--deep" test cases rely on to keep their fixtures
-// deterministic. --deep now only fills in flags the user didn't set explicitly, mirroring how
-// --shallow already respects an explicit --include-*=true. This is a pure viper/pflag check, no
-// live cluster or even a real render involved, so it runs unconditionally.
-func TestDeepRespectsExplicitIncludeFalse(t *testing.T) {
-	v := viper.New()
-	flags := pflag.NewFlagSet("test", pflag.ContinueOnError)
-	addRenderFlags(flags)
-	require.NoError(t, v.BindPFlags(flags))
-	require.NoError(t, flags.Parse([]string{"--deep", "--include-events=false", "--include-managed-fields=false"}))
-
-	require.True(t, v.GetBool("deep"))
-	enableAllIncludes(v)
-
-	assert.False(t, v.GetBool("include-events"), "explicit --include-events=false must survive --deep")
-	assert.False(t, v.GetBool("include-managed-fields"), "explicit --include-managed-fields=false must survive --deep")
-	assert.True(t, v.GetBool("include-owners"), "--deep must still enable flags the user didn't set explicitly")
-}
-
-// TestE2ERegexFixturesAreAnchored guards the whole-output convention documented in
-// CONTRIBUTING.md: a fixture under tests/e2e-artifacts/ either pins the full rendered output,
-// anchored at both ends with `\A`...`\z`, or is a deliberately partial one-off-lines match with
-// neither anchor. A fixture with only one of the two anchors is neither -- almost always a fixture
-// that was meant to be anchored but lost its `\z` (or gained a stray `\A`) while being edited, so
-// it silently stopped verifying the parts of the output past/before the missing anchor. This
-// doesn't require a live cluster, so it runs unconditionally.
-func TestE2ERegexFixturesAreAnchored(t *testing.T) {
-	fixtures, err := filepath.Glob("../tests/e2e-artifacts/*.regex")
-	require.NoError(t, err)
-	require.NotEmpty(t, fixtures)
-	for _, fixture := range fixtures {
-		fixture := fixture
-		t.Run(filepath.Base(fixture), func(t *testing.T) {
-			content, err := os.ReadFile(fixture)
-			require.NoError(t, err)
-			startsAnchored := bytes.HasPrefix(content, []byte(`\A`))
-			endsAnchored := bytes.HasSuffix(content, []byte(`\z`))
-			if startsAnchored != endsAnchored {
-				t.Errorf("%s: has `\\A` at the start=%v but `\\z` at the end=%v -- a whole-output "+
-					"fixture needs both anchors (see CONTRIBUTING.md), a partial one-off-lines "+
-					"fixture needs neither; use the --include-* flags to trim sections you don't "+
-					"want to pin instead of matching only part of the output",
-					fixture, startsAnchored, endsAnchored)
-			}
-		})
-	}
-}
-
-func TestE2EAgainstVanillaMinikube(t *testing.T) {
-	e2eMinikubeTest(t)
-	hackOpts := testHackOpts(t)
-	klog.InitFlags(nil)
-	t.Log("starting tests...")
-	tests := []cmdTest{
-		{
-			name:        "empty call should print an error and usage",
-			stderrRegex: `You must provide one or more resources by argument or filename.\nExample resource specifications`,
-		},
-		{
-			name:        "pods on non-existent ns should fail",
-			args:        []string{"pods", "-n", "non-existent"},
-			stderrRegex: `error: no resources found\n$`,
-		},
-		{
-			name:            "pods on kube-system ns should return storage-provisioner",
-			args:            []string{"pods", "-n", "kube-system", "--include-events=false", "--include-managed-fields=false"},
-			stdoutRegexPath: "e2e-artifacts/pods-kube-system.regex",
-		},
-		{
-			name:            "node query should return at least a node",
-			args:            []string{"node", "--include-events=false", "--include-managed-fields=false"},
-			stdoutRegexPath: "e2e-artifacts/node-query.regex",
-		},
-		{
-			name:        "cr file without a crd should fail",
-			args:        []string{"-f", "../tests/artifacts/cr-dbconn-mymysql.yaml"},
-			stderrRegex: `no matches for kind "DatabaseConnection" in version "example.com/v1alpha1"`,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			tt.assert(t, nil, combineOpts(hackOpts, viperTestHackOpts())...)
-		})
-	}
-}
-
-// testHackOpts fixes plugin.RenderConfig's Now/DurationRound/StartedAfterClause for
-// deterministic e2e output (see plugin.ApplyTestHack). Each RootCmd() invocation gets its own
-// fresh RenderConfig (see cmd/main.go), so unlike the old global package-var overrides this needs
-// no revert -- see #694.
-func testHackOpts(t *testing.T) []func(*plugin.RenderConfig) {
-	t.Helper()
-	return []func(*plugin.RenderConfig){plugin.ApplyTestHack}
-}
-
-// viperTestHackOpts sets "test-hack" on this invocation's RenderConfig, which makes ip() report a
-// fixed 1.1.1.1 instead of the real address.
-func viperTestHackOpts() []func(*plugin.RenderConfig) {
-	return []func(*plugin.RenderConfig){
-		func(cfg *plugin.RenderConfig) {
-			cfg.Viper.Set("test-hack", true)
-		},
-	}
-}
-
-// combineOpts concatenates RenderConfig option groups (e.g. testHackOpts, viperTestHackOpts) into
-// a single slice, applied in order by RootCmd.
-func combineOpts(groups ...[]func(*plugin.RenderConfig)) []func(*plugin.RenderConfig) {
-	var opts []func(*plugin.RenderConfig)
-	for _, g := range groups {
-		opts = append(opts, g...)
-	}
-	return opts
-}
-
-func TestAllArtifactsLocal(t *testing.T) {
-	t.Setenv("KUBECONFIG", "/dev/null")
-	opts := combineOpts(testHackOpts(t), viperTestHackOpts())
-	artifacts, err := filepath.Glob("../tests/artifacts/*.yaml")
-	assert.NoError(t, err)
-	for _, artifact := range artifacts {
-		name := strings.Replace(artifact, "../tests/", "", 1)
-		name = strings.Replace(name, ".yaml", "", 1)
-		t.Run(name, func(t *testing.T) {
-			test := cmdTest{
-				args:            []string{"-f", artifact, "--local", "--shallow", "--v", "255"},
-				stdoutEqualPath: name + ".out",
-			}
-			test.assert(t, nil, opts...) // to update the out files check /tests/artifacts/README.md
-		})
-	}
-}
-
-func TestAllArtifactsLocalWithIncludeAllVolumes(t *testing.T) {
-	t.Setenv("KUBECONFIG", "/dev/null")
-	opts := combineOpts(testHackOpts(t), viperTestHackOpts())
-	artifacts := []string{
-		"../tests/artifacts/pod-standalone.yaml",
-		"../tests/artifacts/pod-missing-pvc.yaml",
-	}
-	for _, artifact := range artifacts {
-		artifact := artifact
-		name := strings.Replace(artifact, "../tests/", "", 1)
-		name = strings.Replace(name, ".yaml", "", 1)
-		t.Run(name, func(t *testing.T) {
-			test := cmdTest{
-				args:            []string{"-f", artifact, "--local", "--shallow", "--v", "255", "--include-all-volumes"},
-				stdoutEqualPath: name + ".include-all-volumes.out",
-			}
-			test.assert(t, nil, opts...)
-		})
-	}
-}
-
-func TestAllArtifactsLocalWithAbsoluteTime(t *testing.T) {
-	t.Setenv("KUBECONFIG", "/dev/null")
-	opts := combineOpts(viperTestHackOpts())
-	artifacts := []string{
-		"../tests/artifacts/pod-standalone.yaml",
-	}
-	for _, artifact := range artifacts {
-		artifact := artifact
-		name := strings.Replace(artifact, "../tests/", "", 1)
-		name = strings.Replace(name, ".yaml", "", 1)
-		t.Run(name, func(t *testing.T) {
-			test := cmdTest{
-				args:            []string{"-f", artifact, "--local", "--shallow", "--v", "255", "--absolute-time"},
-				stdoutEqualPath: name + ".absolute-time.out",
-			}
-			test.assert(t, nil, opts...)
-		})
-	}
-}
-
-func executeCMD(t *testing.T, args []string, opts ...func(*plugin.RenderConfig)) (string, string, error) {
-	t.Helper()
-	cmd := RootCmd(opts...)
-	stdout := &bytes.Buffer{}
-	cmd.SetOut(stdout)
-	stderr := &bytes.Buffer{}
-	cmd.SetErr(stderr)
-	cmd.SetArgs(args)
-	t.Logf("running command with: %s", strings.Join(args, " "))
-	err := cmd.Execute()
-	return stdout.String(), stderr.String(), err
-}
-
-func startMinikube(t *testing.T) {
-	t.Helper()
-	// `make test-e2e` computes an isolated profile name (branch + Claude Code session
-	// hash, see Makefile) and passes ASSUME_MINIKUBE_IS_CONFIGURED=true, so it never
-	// reaches this function. This fallback only matters for ad hoc
-	// `go test -run TestE2E...` invocations that bypass the Makefile: set E2E_PROFILE
-	// yourself (`make print-e2e-profile` prints the same name the Makefile would use)
-	// to avoid colliding with other worktrees/sessions on a shared t.Name() profile.
-	// TODO: derive branch+session identity here directly instead of relying on the
-	// caller to export E2E_PROFILE, so ad hoc runs are isolated automatically too.
-	clusterName := os.Getenv("E2E_PROFILE")
-	if clusterName == "" {
-		clusterName = t.Name()
-	}
-	t.Logf("Creating temp folder for minikube.kubeconfig for minikube %s ...", clusterName)
-	dir, err := os.MkdirTemp("", clusterName)
-	require.NoError(t, err)
-	kubeconfig := path.Join(dir, "minikube.kubeconfig")
-	t.Setenv("KUBECONFIG", kubeconfig)
-	t.Logf("Starting Minikube cluster %s with %s ...", clusterName, kubeconfig)
-	// --cpus/--memory: matches the Makefile's e2e-minikube-up sizing, needed for TestE2EParallel's
-	// subtests to run concurrently without overwhelming the VM (see that target's comment).
-	startMinikube := exec.Command("minikube", "start", "-p", clusterName, "--addons=metrics-server", "--cpus=4", "--memory=6g")
-	require.NoError(t, startMinikube.Run())
-	require.NoError(t, exec.Command("kubectl", "-n", "kube-system", "rollout", "status",
-		"deployment/metrics-server", "--timeout=120s").Run())
-	t.Cleanup(func() {
-		cmd := exec.Command("minikube", "delete", "-p", clusterName)
-		t.Logf("Deleting Minikube cluster %s...", clusterName)
-		if err := cmd.Run(); err != nil {
-			t.Log("Error deleting Minikube cluster:", err)
-		}
-		t.Logf("Deleting temp folder for minikube %s: %s ...", clusterName, dir)
-		if err := os.RemoveAll(dir); err != nil {
-			t.Log("Error deleting temp folder of minikube.kubeconfig:", err)
-		}
-	})
-}
-
-func e2eMinikubeTest(t *testing.T) {
-	t.Helper()
-	if os.Getenv("RUN_E2E_TESTS") != "true" {
-		t.Skip("Skipping e2e test as RUN_E2E_TESTS is not set to true")
-	}
-	if os.Getenv("ASSUME_MINIKUBE_IS_CONFIGURED") == "true" {
-		t.Logf("assuming current kubeconfig context is pointng a minikube to run e2e tests")
-	} else {
-		startMinikube(t)
-	}
-}
-
-func e2eClients(t *testing.T) ([]func(*plugin.RenderConfig), *kubernetes.Clientset, dynamic.Interface) {
-	t.Helper()
-	hackOpts := testHackOpts(t)
-	kubeconfigPath := os.Getenv("KUBECONFIG")
-	if kubeconfigPath == "" {
-		homeDir, err := os.UserHomeDir()
-		if err != nil {
-			t.Fatalf("failed to get user home directory: %v", err)
-		}
-		kubeconfigPath = filepath.Join(homeDir, ".kube", "config")
-	}
-	config, err := clientcmd.BuildConfigFromFlags("", kubeconfigPath)
-	if err != nil {
-		t.Fatal(err)
-	}
-	clientset, err := kubernetes.NewForConfig(config)
-	if err != nil {
-		t.Fatal(err)
-	}
-	dynamicClient, err := dynamic.NewForConfig(config)
-	if err != nil {
-		t.Fatal(err)
-	}
-	return hackOpts, clientset, dynamicClient
-}
 
 // TestE2EParallel is a dedicated home for e2e subtests that are independent of each other and can
 // therefore run concurrently. RootCmd (cmd/main.go) and pkg/plugin no longer read a process-global
@@ -1923,498 +1490,4 @@ func TestE2EParallel(t *testing.T) {
 			stdoutRegexPath: "e2e-artifacts/node-metrics-multi-namespace.regex",
 		}.assert(t, nil, opts...)
 	})
-}
-
-func TestE2EDynamicManifests(t *testing.T) {
-	e2eMinikubeTest(t)
-	hackOpts, clientset, dynamicClient := e2eClients(t)
-	t.Run("pod containers section warns when metrics-server's APIService is missing", func(t *testing.T) {
-		// Issue #165 case 1: metrics-server was never installed. We simulate that by removing
-		// just the APIService object that fronts it (not the Deployment/Service), which is
-		// exactly what KubeMetricsUnavailableReason checks -- so the round trip is near-instant
-		// and doesn't disturb metrics-server's actual health for other subtests.
-		opts := combineOpts(hackOpts, viperTestHackOpts())
-		apiServiceYAML, err := exec.Command("kubectl", "get", "apiservice", "v1beta1.metrics.k8s.io", "-o", "yaml").Output()
-		require.NoError(t, err)
-		require.NoError(t, exec.Command("kubectl", "delete", "apiservice", "v1beta1.metrics.k8s.io").Run())
-		t.Cleanup(func() {
-			applyCmd := exec.Command("kubectl", "apply", "-f", "-")
-			applyCmd.Stdin = bytes.NewReader(apiServiceYAML)
-			require.NoError(t, applyCmd.Run())
-			waitForMetricsAPIServiceAvailable(t)
-		})
-
-		_, err = clientset.CoreV1().Pods("default").Create(context.TODO(), &corev1.Pod{
-			ObjectMeta: metav1.ObjectMeta{Name: "e2e-pod-metrics-server-missing"},
-			Spec: corev1.PodSpec{
-				Containers: []corev1.Container{{Name: "main", Image: "busybox", Command: []string{"sleep", "infinity"}}},
-			},
-		}, metav1.CreateOptions{})
-		require.NoError(t, err)
-		t.Cleanup(func() {
-			clientset.CoreV1().Pods("default").Delete(context.TODO(), "e2e-pod-metrics-server-missing", metav1.DeleteOptions{})
-		})
-		require.NoError(t, exec.Command("kubectl", "wait", "--for=condition=Ready",
-			"pod/e2e-pod-metrics-server-missing", "--timeout=2m").Run())
-
-		cmdTest{
-			args:            []string{"pod/e2e-pod-metrics-server-missing", "--include-events=false", "--include-managed-fields=false", "--v", "5"},
-			stdoutRegexPath: "e2e-artifacts/pod-metrics-server-missing.regex",
-		}.assert(t, nil, opts...)
-	})
-	t.Run("VerticalPodAutoscaler reverse-matches its target workload and shows an applied recommendation", func(t *testing.T) {
-		// Deliberately kept out of TestE2EParallel's pool: the burner container below
-		// intentionally pegs a full CPU to give the VPA recommender a reason to act, and on a
-		// single-node cluster that starves metrics-server's own readiness probe when it runs
-		// alongside the other concurrent subtests -- causing unrelated renders elsewhere to
-		// intermittently report "metrics-server is not available". Running it serially, alongside
-		// the other genuinely cluster-wide-affecting subtest above, avoids that.
-		opts := combineOpts(hackOpts, viperTestHackOpts())
-		ns := "e2e-vpa"
-		_, err := clientset.CoreV1().Namespaces().Create(context.TODO(),
-			&corev1.Namespace{ObjectMeta: metav1.ObjectMeta{Name: ns}}, metav1.CreateOptions{})
-		require.NoError(t, err)
-		t.Cleanup(func() {
-			clientset.CoreV1().Namespaces().Delete(context.TODO(), ns, metav1.DeleteOptions{})
-		})
-
-		name := "vpa-burner"
-		one := int32(1)
-		dep := &appsv1.Deployment{
-			ObjectMeta: metav1.ObjectMeta{Name: name, Namespace: ns},
-			Spec: appsv1.DeploymentSpec{
-				Replicas: &one,
-				Selector: &metav1.LabelSelector{MatchLabels: map[string]string{"app": name}},
-				Template: corev1.PodTemplateSpec{
-					ObjectMeta: metav1.ObjectMeta{Labels: map[string]string{"app": name}},
-					Spec: corev1.PodSpec{Containers: []corev1.Container{{
-						Name:    "burner",
-						Image:   "busybox",
-						Command: []string{"sh", "-c", "yes > /dev/null"},
-						Resources: corev1.ResourceRequirements{
-							Requests: corev1.ResourceList{
-								corev1.ResourceCPU:    resource.MustParse("10m"),
-								corev1.ResourceMemory: resource.MustParse("16Mi"),
-							},
-						},
-					}}},
-				},
-			},
-		}
-		_, err = clientset.AppsV1().Deployments(ns).Create(context.TODO(), dep, metav1.CreateOptions{})
-		require.NoError(t, err)
-		require.NoError(t, exec.Command("kubectl", "wait", "--for=condition=Available",
-			"deployment/"+name, "-n", ns, "--timeout=4m").Run())
-		originalPod := waitForPodByLabel(t, ns, "app="+name)
-
-		vpaGVR := schema.GroupVersionResource{Group: "autoscaling.k8s.io", Version: "v1", Resource: "verticalpodautoscalers"}
-		vpa := &unstructured.Unstructured{Object: map[string]interface{}{
-			"apiVersion": "autoscaling.k8s.io/v1",
-			"kind":       "VerticalPodAutoscaler",
-			"metadata":   map[string]interface{}{"name": name, "namespace": ns},
-			"spec": map[string]interface{}{
-				"targetRef": map[string]interface{}{"apiVersion": "apps/v1", "kind": "Deployment", "name": name},
-				"updatePolicy": map[string]interface{}{
-					"updateMode":  "Recreate",
-					"minReplicas": int64(1),
-				},
-				"resourcePolicy": map[string]interface{}{
-					"containerPolicies": []interface{}{
-						map[string]interface{}{
-							"containerName": "burner",
-							"minAllowed":    map[string]interface{}{"cpu": "10m", "memory": "16Mi"},
-							"maxAllowed":    map[string]interface{}{"cpu": "500m", "memory": "128Mi"},
-						},
-					},
-				},
-			},
-		}}
-		_, err = dynamicClient.Resource(vpaGVR).Namespace(ns).Create(context.TODO(), vpa, metav1.CreateOptions{})
-		require.NoError(t, err)
-		defer dynamicClient.Resource(vpaGVR).Namespace(ns).Delete(context.TODO(), name, metav1.DeleteOptions{})
-
-		waitForVPARecommendation(t, ns, name)
-		waitForPodRecreated(t, ns, "app="+name, originalPod)
-		// The evicted Pod can briefly still be listed (Terminating) alongside the replacement --
-		// wait for exactly one to remain so the fixture below can pin a single Pod line.
-		waitForSinglePod(t, ns, "app="+name)
-		// waitForPodRecreated/waitForSinglePod only check the replacement Pod's name/count, not
-		// its readiness -- under concurrent cluster load its Running/Ready transition can lag
-		// well behind that, and the fixture below pins the Deployment as fully Available, so wait
-		// for that explicitly rather than racing the kubelet.
-		require.NoError(t, exec.Command("kubectl", "wait", "--for=condition=Available",
-			"deployment/"+name, "-n", ns, "--timeout=5m").Run())
-		waitForVPAPodsMatched(t, ns, name)
-
-		cmdTest{
-			args:            []string{"deployment/" + name, "-n", ns, "--include-events=false", "--include-managed-fields=false", "--v", "5"},
-			stdoutRegexPath: "e2e-artifacts/vpa-workload-reverse-match.regex",
-		}.assert(t, nil, opts...)
-		cmdTest{
-			args:            []string{"vpa/" + name, "-n", ns, "--include-managed-fields=false", "--v", "5"},
-			stdoutRegexPath: "e2e-artifacts/vpa-standalone.regex",
-		}.assert(t, nil, opts...)
-	})
-	t.Run("Crossplane XR composes namespaced children and surfaces their health", func(t *testing.T) {
-		// Crossplane core plus the two Composition Functions it needs (installed cluster-wide by
-		// `make install-e2e-deps`) must actually reconcile to produce the XR's composed children,
-		// same "controller must actually run" reasoning as the VPA subtest above.
-		opts := combineOpts(hackOpts, viperTestHackOpts())
-		ns := "e2e-crossplane-xr"
-		_, err := clientset.CoreV1().Namespaces().Create(context.TODO(),
-			&corev1.Namespace{ObjectMeta: metav1.ObjectMeta{Name: ns}}, metav1.CreateOptions{})
-		require.NoError(t, err)
-		t.Cleanup(func() {
-			clientset.CoreV1().Namespaces().Delete(context.TODO(), ns, metav1.DeleteOptions{})
-		})
-
-		applyManifest(t, "e2e-artifacts/crossplane-xstatusprobe.yaml")
-		require.NoError(t, exec.Command("kubectl", "wait", "--for=condition=Established",
-			"xrd/xstatusprobes.tests.kubectl-status.io", "--timeout=60s").Run())
-		applyManifestInNamespace(t, "e2e-artifacts/crossplane-xr.yaml", ns)
-		waitForInNamespace(t, "xstatusprobe/probe-a", "condition=Synced", ns)
-		// The Deployment child is deliberately unschedulable (a nodeSelector no node can match),
-		// so the XR itself never reaches Ready -- wait on the field kubectl-status actually reads
-		// instead of a condition that will never flip.
-		waitForCrossplaneComposedRefs(t, ns, "probe-a", 2)
-		// Synced/resourceRefs land as soon as the render step runs, but the XR's own Responsive
-		// condition and the composed Deployment's Progressing/Available conditions populate
-		// slightly later via separate reconciles -- wait for all of them so the fixtures below
-		// pin a stable message instead of racing a transient "Replicas: 0/1" kstatus summary.
-		waitForInNamespace(t, "xstatusprobe/probe-a", "condition=Responsive", ns)
-		waitForInNamespace(t, "deployment/probe-a-blocked", "condition=Progressing", ns)
-		require.NoError(t, exec.Command("kubectl", "wait", "-n", ns,
-			"--for=condition=PodScheduled=false", "pod", "-l", "app=probe-a-blocked", "--timeout=2m").Run())
-		// kstatus (sigs.k8s.io/cli-utils/pkg/kstatus/status.ScheduleWindow) gives a Pod 15s from
-		// its creationTimestamp before reporting Unschedulable as Failed rather than InProgress --
-		// wait that out so the fixtures below pin the stable "Failed: Pod could not be scheduled"
-		// message instead of racing the transient one.
-		waitForPodScheduleWindow(t, ns, "app=probe-a-blocked")
-
-		// Only the live-query-dependent branches belong here: default mode's KubeGetFirst lookup
-		// (populating each composed child's compact health) and --deep's IncludeRenderableObject
-		// inline. Shallow rendering and Composition.tmpl make no live queries at all -- both are
-		// already covered by the offline artifacts (tests/artifacts/crossplane-*).
-		cmdTest{
-			args:            []string{"xstatusprobe/probe-a", "-n", ns, "--include-events=false", "--include-managed-fields=false", "--v", "5"},
-			stdoutRegexPath: "e2e-artifacts/crossplane-xr.regex",
-		}.assert(t, nil, opts...)
-		cmdTest{
-			args:            []string{"xstatusprobe/probe-a", "-n", ns, "--deep", "--include-events=false", "--include-managed-fields=false", "--v", "5"},
-			stdoutRegexPath: "e2e-artifacts/crossplane-xr-deep.regex",
-		}.assert(t, nil, opts...)
-	})
-}
-
-// waitForPodScheduleWindow blocks until at least 15s (kstatus's
-// sigs.k8s.io/cli-utils/pkg/kstatus/status.ScheduleWindow) have passed since the matching Pod's
-// creationTimestamp.
-func waitForPodScheduleWindow(t *testing.T, namespace, labelSelector string) {
-	t.Helper()
-	cmd := exec.Command("kubectl", "get", "pods", "-n", namespace, "-l", labelSelector,
-		"-o", "jsonpath={.items[0].metadata.creationTimestamp}")
-	output, err := cmd.CombinedOutput()
-	require.NoError(t, err)
-	created, err := time.Parse(time.RFC3339, strings.TrimSpace(string(output)))
-	require.NoError(t, err)
-	if remaining := time.Until(created.Add(16 * time.Second)); remaining > 0 {
-		time.Sleep(remaining)
-	}
-}
-
-// waitForCrossplaneComposedRefs polls until the XR's spec.crossplane.resourceRefs has at least
-// wantCount entries. Used instead of waiting on a Ready condition since the XR under test never
-// reaches Ready (one composed child is deliberately unschedulable).
-func waitForCrossplaneComposedRefs(t *testing.T, namespace, name string, wantCount int) {
-	t.Helper()
-	deadline := time.Now().Add(4 * time.Minute)
-	for time.Now().Before(deadline) {
-		cmd := exec.Command("kubectl", "get", "xstatusprobe", name, "-n", namespace,
-			"-o", "jsonpath={.spec.crossplane.resourceRefs}")
-		output, err := cmd.CombinedOutput()
-		if err == nil {
-			var refs []interface{}
-			if json.Unmarshal(output, &refs) == nil && len(refs) >= wantCount {
-				t.Logf("xstatusprobe %s in namespace %s has %d composed resource refs", name, namespace, len(refs))
-				return
-			}
-		}
-		time.Sleep(2 * time.Second)
-	}
-	t.Fatalf("timed out waiting for xstatusprobe %s in namespace %s to have %d composed resource refs", name, namespace, wantCount)
-}
-
-func applyManifest(t *testing.T, filepath string) {
-	t.Helper()
-	filepath = path.Join("..", "tests", filepath)
-	cmd := exec.Command("kubectl", "apply", "-f", filepath)
-	output, err := cmd.CombinedOutput()
-	t.Cleanup(func() {
-		t.Logf("deleting manifest %s", filepath)
-		cmd := exec.Command("kubectl", "delete", "-f", filepath)
-		output, err := cmd.CombinedOutput()
-		if err != nil {
-			t.Logf("warning: failed to delete manifest %s: %v (output: %s)", filepath, err, string(output))
-			return
-		}
-		t.Logf("manifest deleted %s: %s", filepath, string(output))
-	})
-	require.NoError(t, err)
-	t.Logf("applied manifest %s: %s", filepath, string(output))
-}
-
-// applyManifestInNamespace is applyManifest, but targets a namespace via `kubectl -n` instead of
-// relying on the manifest's own metadata.namespace (or the kubeconfig's default) -- used to give a
-// subtest a dedicated namespace without needing a namespace-specific copy of its fixture yaml. The
-// manifest's objects must not already set their own metadata.namespace, since that always wins
-// over `-n` and would silently defeat the isolation this is for.
-func applyManifestInNamespace(t *testing.T, filepath, namespace string) {
-	t.Helper()
-	filepath = path.Join("..", "tests", filepath)
-	cmd := exec.Command("kubectl", "apply", "-n", namespace, "-f", filepath)
-	output, err := cmd.CombinedOutput()
-	t.Cleanup(func() {
-		t.Logf("deleting manifest %s from namespace %s", filepath, namespace)
-		cmd := exec.Command("kubectl", "delete", "-n", namespace, "-f", filepath)
-		output, err := cmd.CombinedOutput()
-		if err != nil {
-			t.Logf("warning: failed to delete manifest %s from namespace %s: %v (output: %s)", filepath, namespace, err, string(output))
-			return
-		}
-		t.Logf("manifest deleted %s from namespace %s: %s", filepath, namespace, string(output))
-	})
-	require.NoError(t, err)
-	t.Logf("applied manifest %s to namespace %s: %s", filepath, namespace, string(output))
-}
-
-// waitForInNamespace targets a namespace explicitly via `kubectl -n` instead of the kubeconfig's
-// default -- pairs with applyManifestInNamespace for subtests moved off the shared default
-// namespace.
-func waitForInNamespace(t *testing.T, resource, forParam, namespace string) {
-	t.Helper()
-	cmd := exec.Command("kubectl", "wait", "-n", namespace, "--for", forParam, resource, "--timeout=4m")
-	output, err := cmd.CombinedOutput()
-	t.Logf("wait result for %s in namespace %s: %s", resource, namespace, string(output))
-	require.NoError(t, err)
-}
-
-// waitForContainerRestart polls until the named container in the resource reports a
-// restartCount greater than zero.
-// waitForSinglePod polls until exactly one pod matches the given label selector. Used after a
-// rollout to make sure the previous revision's pod has actually finished terminating: `kubectl
-// rollout status` and the Deployment's `.status.replicas` field can both report the rollout as
-// done slightly before the old pod object is removed, which otherwise makes the rendered output
-// briefly list two Pods instead of one.
-func waitForSinglePod(t *testing.T, namespace, labelSelector string) {
-	t.Helper()
-	deadline := time.Now().Add(4 * time.Minute)
-	for time.Now().Before(deadline) {
-		cmd := exec.Command("kubectl", "get", "pods", "-n", namespace, "-l", labelSelector,
-			"-o", "jsonpath={.items[*].metadata.name}")
-		output, err := cmd.CombinedOutput()
-		if err == nil {
-			names := strings.Fields(string(output))
-			if len(names) == 1 {
-				t.Logf("exactly one pod %s matches selector %s in namespace %s", names[0], labelSelector, namespace)
-				return
-			}
-		}
-		time.Sleep(2 * time.Second)
-	}
-	t.Fatalf("timed out waiting for exactly one pod matching selector %s in namespace %s", labelSelector, namespace)
-}
-
-// waitForContainerWaitingReason polls until the named container in the resource reports the
-// given waiting-state reason. Used instead of a plain restart-count check because a crashlooping
-// container's current state flips between Waiting(CrashLoopBackOff) and Terminated(Error) as the
-// kubelet retries, so waiting for a stable, specific state avoids a flaky render.
-// waitForPodByLabel polls until exactly one pod matches the given label selector and returns
-// its name. Used for Deployment/DaemonSet, whose pod names include a random suffix that isn't
-// known ahead of time (unlike StatefulSet, where pod names are predictable).
-func waitForPodByLabel(t *testing.T, namespace, labelSelector string) string {
-	t.Helper()
-	deadline := time.Now().Add(4 * time.Minute)
-	for time.Now().Before(deadline) {
-		cmd := exec.Command("kubectl", "get", "pods", "-n", namespace, "-l", labelSelector,
-			"-o", "jsonpath={.items[*].metadata.name}")
-		output, err := cmd.CombinedOutput()
-		if err == nil {
-			names := strings.Fields(string(output))
-			if len(names) == 1 {
-				name := names[0]
-				t.Logf("found pod %s matching selector %s in namespace %s", name, labelSelector, namespace)
-				return name
-			}
-		}
-		time.Sleep(2 * time.Second)
-	}
-	t.Fatalf("timed out waiting for exactly one pod matching selector %s in namespace %s", labelSelector, namespace)
-	return ""
-}
-
-// waitForContainerWaitingReasonInNamespace targets a namespace explicitly via `kubectl -n`
-// instead of the kubeconfig's default; pass "" to use the kubeconfig's default namespace.
-func waitForContainerWaitingReasonInNamespace(t *testing.T, resource, containerName, reason, namespace string) {
-	t.Helper()
-	jsonpath := fmt.Sprintf(`{.status.containerStatuses[?(@.name=="%s")].state.waiting.reason}`, containerName)
-	args := []string{"get", resource, "-o", "jsonpath=" + jsonpath}
-	if namespace != "" {
-		args = append([]string{"-n", namespace}, args...)
-	}
-	deadline := time.Now().Add(4 * time.Minute)
-	for time.Now().Before(deadline) {
-		cmd := exec.Command("kubectl", args...)
-		output, err := cmd.CombinedOutput()
-		if err == nil && strings.TrimSpace(string(output)) == reason {
-			t.Logf("%s container %s reached waiting reason %s", resource, containerName, reason)
-			return
-		}
-		time.Sleep(2 * time.Second)
-	}
-	t.Fatalf("timed out waiting for %s container %s to report waiting reason %s", resource, containerName, reason)
-}
-
-// waitForPodMetrics polls the metrics.k8s.io API directly until it has scraped data for the
-// given pod. metrics-server's scrape interval means a freshly-created pod's metrics aren't
-// available immediately after it goes Ready.
-func waitForPodMetrics(t *testing.T, namespace, name string) {
-	t.Helper()
-	rawPath := fmt.Sprintf("/apis/metrics.k8s.io/v1beta1/namespaces/%s/pods/%s", namespace, name)
-	deadline := time.Now().Add(4 * time.Minute)
-	for time.Now().Before(deadline) {
-		if err := exec.Command("kubectl", "get", "--raw", rawPath).Run(); err == nil {
-			t.Logf("metrics available for pod %s/%s", namespace, name)
-			return
-		}
-		time.Sleep(2 * time.Second)
-	}
-	t.Fatalf("timed out waiting for metrics.k8s.io data for pod %s/%s", namespace, name)
-}
-
-// waitForContainerMetrics polls metrics.k8s.io until every named container has a recorded cpu
-// usage. A pod's PodMetrics can exist while still missing an entry for a container that started
-// slightly later than its siblings (more likely under concurrent cluster load): Pod.tmpl's
-// container_status_summary only renders a usage line once a container's own entry is present
-// with usage.cpu set, so asserting before that leaves the container's line silently blank
-// instead of matching a golden fixture.
-func waitForContainerMetrics(t *testing.T, namespace, name string, containerNames ...string) {
-	t.Helper()
-	rawPath := fmt.Sprintf("/apis/metrics.k8s.io/v1beta1/namespaces/%s/pods/%s", namespace, name)
-	type containerMetrics struct {
-		Name  string `json:"name"`
-		Usage struct {
-			CPU string `json:"cpu"`
-		} `json:"usage"`
-	}
-	type podMetrics struct {
-		Containers []containerMetrics `json:"containers"`
-	}
-	deadline := time.Now().Add(4 * time.Minute)
-	for time.Now().Before(deadline) {
-		output, err := exec.Command("kubectl", "get", "--raw", rawPath).Output()
-		if err == nil {
-			var m podMetrics
-			if json.Unmarshal(output, &m) == nil {
-				have := map[string]bool{}
-				for _, c := range m.Containers {
-					if c.Usage.CPU != "" {
-						have[c.Name] = true
-					}
-				}
-				allPresent := true
-				for _, want := range containerNames {
-					if !have[want] {
-						allPresent = false
-						break
-					}
-				}
-				if allPresent {
-					t.Logf("metrics available for all of %v in pod %s/%s", containerNames, namespace, name)
-					return
-				}
-			}
-		}
-		time.Sleep(2 * time.Second)
-	}
-	t.Fatalf("timed out waiting for metrics.k8s.io cpu usage for containers %v in pod %s/%s", containerNames, namespace, name)
-}
-
-// waitForMetricsAPIServiceAvailable polls until the metrics-server APIService reports
-// Available=True. Used after recreating it post-deletion: the backing Deployment/Service were
-// never touched, so this is a quick re-sync, not the ~1 minute metrics-server itself needs to
-// scrape fresh data.
-func waitForMetricsAPIServiceAvailable(t *testing.T) {
-	t.Helper()
-	deadline := time.Now().Add(2 * time.Minute)
-	for time.Now().Before(deadline) {
-		output, err := exec.Command("kubectl", "get", "apiservice", "v1beta1.metrics.k8s.io",
-			"-o", `jsonpath={.status.conditions[?(@.type=="Available")].status}`).Output()
-		if err == nil && strings.TrimSpace(string(output)) == "True" {
-			t.Log("metrics-server APIService is Available again")
-			return
-		}
-		time.Sleep(2 * time.Second)
-	}
-	t.Fatalf("timed out waiting for metrics-server APIService to become Available again")
-}
-
-// waitForVPARecommendation polls until a VerticalPodAutoscaler's status.recommendation is
-// populated. The recommender needs a window of real usage samples before it computes a first
-// recommendation, so this can take roughly a minute after the VPA and its target Pod both exist.
-func waitForVPARecommendation(t *testing.T, namespace, name string) {
-	t.Helper()
-	deadline := time.Now().Add(5 * time.Minute)
-	for time.Now().Before(deadline) {
-		output, err := exec.Command("kubectl", "get", "vpa", name, "-n", namespace,
-			"-o", "jsonpath={.status.recommendation.containerRecommendations[0].target.cpu}").CombinedOutput()
-		if err == nil && strings.TrimSpace(string(output)) != "" {
-			t.Logf("VPA %s/%s has a recommendation: %s", namespace, name, strings.TrimSpace(string(output)))
-			return
-		}
-		time.Sleep(3 * time.Second)
-	}
-	t.Fatalf("timed out waiting for VPA %s/%s to compute a recommendation", namespace, name)
-}
-
-// waitForVPAPodsMatched polls until a VerticalPodAutoscaler's NoPodsMatched condition is gone
-// (or already False). The VPA controller re-evaluates this independently of, and can lag behind,
-// the target Deployment's own Available condition -- right after the updater recreates the Pod,
-// the VPA can still be reporting stale NoPodsMatched=True for a beat, which the golden fixture
-// doesn't expect to see at all.
-func waitForVPAPodsMatched(t *testing.T, namespace, name string) {
-	t.Helper()
-	deadline := time.Now().Add(2 * time.Minute)
-	for time.Now().Before(deadline) {
-		output, err := exec.Command("kubectl", "get", "vpa", name, "-n", namespace,
-			"-o", `jsonpath={.status.conditions[?(@.type=="NoPodsMatched")].status}`).CombinedOutput()
-		if err == nil && strings.TrimSpace(string(output)) != "True" {
-			t.Logf("VPA %s/%s no longer reports NoPodsMatched", namespace, name)
-			return
-		}
-		time.Sleep(2 * time.Second)
-	}
-	t.Fatalf("timed out waiting for VPA %s/%s to clear its NoPodsMatched condition", namespace, name)
-}
-
-// waitForPodRecreated polls until the Pod matching labelSelector is no longer originalPodName --
-// evidence the VPA updater actually evicted/recreated it to apply the recommendation, not just
-// computed one that nobody applied.
-func waitForPodRecreated(t *testing.T, namespace, labelSelector, originalPodName string) {
-	t.Helper()
-	deadline := time.Now().Add(5 * time.Minute)
-	for time.Now().Before(deadline) {
-		output, err := exec.Command("kubectl", "get", "pods", "-n", namespace, "-l", labelSelector,
-			"-o", "jsonpath={.items[*].metadata.name}").CombinedOutput()
-		if err == nil {
-			for _, name := range strings.Fields(string(output)) {
-				if name != originalPodName {
-					t.Logf("VPA updater recreated the pod: %s -> %s", originalPodName, name)
-					return
-				}
-			}
-		}
-		time.Sleep(3 * time.Second)
-	}
-	t.Fatalf("timed out waiting for pod matching %s in namespace %s to be recreated (still %s)",
-		labelSelector, namespace, originalPodName)
 }
